+++ src/vue/vuex/actions.js
@@ -5,10 +5,10 @@ export default {
   changeRoom: ({commit, state}, room) => commit('CHANGE_ROOM', room),
   toggleChatWindow: ({commit, state}) => commit('TOGGLE_CHAT_WINDOW'),
   chatTarget: ({commit}, {email, options = {}}) => {
-    return qiscus.chatTarget(email, options)
+    return QiscusSDK.core.chatTarget(email, options)
     .then((response) => {
       commit('CHAT_TARGET', {email, options})
-      const selected = qiscus.selected.comments
+      const selected = QiscusSDK.core.selected.comments
       const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
       setTimeout(function(){
         if(latestCommentId > 0){
@@ -18,7 +18,7 @@ export default {
         //on entering the room, wait for data processed then focus on comment form
         document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
       }, 0)
-      return Promise.resolve(qiscus.selected)
+      return Promise.resolve(QiscusSDK.core.selected)
     }, (error) => {
       // vm.$toasted.error('Error getting chat room. Please make sure the target is valid')
       return Promise.reject(error)
@@ -26,7 +26,7 @@ export default {
   },
   chatGroup: ({commit}, {id, oldSelected}) => {
     commit('CHAT_GROUP', {id, oldSelected})
-    const selected = qiscus.selected.comments
+    const selected = QiscusSDK.core.selected.comments
     const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
     setTimeout(function(){
       if(latestCommentId > 0){
@@ -39,10 +39,10 @@ export default {
   },
   backToHome: ({commit}) => commit('BACK_TO_HOME'),
   submitComment: ({commit}, {topic_id, comment}) => {
-    return qiscus.submitComment(topic_id, comment)
+    return QiscusSDK.core.submitComment(topic_id, comment)
     .then((response) => {
-      commit('SUBMIT_COMMENT', qiscus.selected)
-      const selected = qiscus.selected.comments
+      commit('SUBMIT_COMMENT', QiscusSDK.core.selected)
+      const selected = QiscusSDK.core.selected.comments
       const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
       setTimeout(function(){
         if(latestCommentId > 0){
@@ -52,16 +52,16 @@ export default {
         //on entering the room, wait for data processed then focus on comment form
         document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
       }, 0)
-      return Promise.resolve(qiscus.selected);
+      return Promise.resolve(QiscusSDK.core.selected);
     }, (error) => {
       return Promise.reject(error)
     })
   },
   submitCommentWithPayload: ({commit}, {topic_id, comment, payload_type, payload}) => {
-    return qiscus.submitComment(topic_id, comment, null, 'reply', JSON.stringify(payload)) 
+    return QiscusSDK.core.submitComment(topic_id, comment, null, 'reply', JSON.stringify(payload)) 
     .then( response => {
-      commit('SUBMIT_COMMENT', qiscus.selected)
-      const selected = qiscus.selected.comments
+      commit('SUBMIT_COMMENT', QiscusSDK.core.selected)
+      const selected = QiscusSDK.core.selected.comments
       const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
       setTimeout(function(){
         if(latestCommentId > 0){
@@ -71,16 +71,16 @@ export default {
         //on entering the room, wait for data processed then focus on comment form
         document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
       }, 0)
-      return Promise.resolve(qiscus.selected);
+      return Promise.resolve(QiscusSDK.core.selected);
     }, (error) => {
       return Promise.reject(error)
     })
   },
   resendComment: ({commit}, comment) => {
-    return qiscus.resendComment(comment)
+    return QiscusSDK.core.resendComment(comment)
     .then((response) => {
-      commit('SUBMIT_COMMENT', qiscus.selected)
-      const selected = qiscus.selected.comments
+      commit('SUBMIT_COMMENT', QiscusSDK.core.selected)
+      const selected = QiscusSDK.core.selected.comments
       const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
       setTimeout(function(){
         if(latestCommentId > 0){
@@ -91,7 +91,7 @@ export default {
         document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
       }, 0)
       console.info('resend comment successful')
-      return Promise.resolve(qiscus.selected);
+      return Promise.resolve(QiscusSDK.core.selected);
     }, error => {
       console.error('resend comment error', error)
       return Promise.reject(error)
@@ -108,7 +108,7 @@ export default {
           ? selectedComment.unique_id === payload.unique_temp_id
           : selectedComment.id === payload.id
       )
-    })(qiscus.selected.comments)
+    })(QiscusSDK.core.selected.comments)
     if (commentToFind){
       commentToFind.markAsRead()
       commit('SET_READ', commentToFind)